Test out a hypothetical option for zlib "deflate" to be applied to the individual records of an FST file, on top of any usual encoding.
To use it for writing records, add 32 to the usual datyp.  Currently, it gets
decoded and stripped out (returns datyp-32) when reading the records back in,
for compatibility with code that does not expect this datyp.

New FST options:
EXTERN_AUTO_ENCODE - automatically encode all output with zlib (off by default)
EXTERN_AUTO_DECODE - automatically decode input and strip datyp (on by default)

In theory, other compression algorithms could be supported as well, but so far
only implemented the support for zlib.