Test out using variable-length codes (specifically Rice codes) for encoding the Lorenzo predictor errors of the compression algorithm.
If the errors follow a geometric distribution, then this should be a more
efficient encoding of the residues.  However, in practice this only yields
about a 3% improvement in efficiency.  The zfstlib routines already seem to be
super efficient, even with the fixed-length codes.  Probably the way it's doing
the local 3x3 tiling is already finding more important efficiency savings.

Recording this proof-of-concept work in a commit just for posterity.